+++ kukumo-plugins/kukumo-rest/src/main/java/iti/kukumo/rest/helpers/TextHelper.java
@@ -10,15 +10,11 @@
 package iti.kukumo.rest.helpers;
 
 
-import iti.kukumo.api.datatypes.Assertion;
-import org.junit.ComparisonFailure;
-
 import io.restassured.http.ContentType;
-import io.restassured.response.ValidatableResponse;
 import iti.commons.jext.Extension;
 import iti.kukumo.rest.ContentTypeHelper;
 import iti.kukumo.rest.MatchMode;
-
+import org.junit.ComparisonFailure;
 
 
 @Extension(provider = "iti.kukumo", name = "rest-text-helper", extensionPoint = "iti.kukumo.rest.ContentTypeHelper")
@@ -34,34 +30,17 @@ public class TextHelper implements ContentTypeHelper {
     @Override
     public void assertContent(String expected, String actual, MatchMode matchMode) {
         switch (matchMode) {
-        case STRICT:
-        case STRICT_ANY_ORDER:
-            if (!actual.trim().equals(expected.trim())) {
-                throw new ComparisonFailure("Text differences", expected, actual);
-            }
-            break;
-        case LOOSE:
-            if (!actual.trim().contains(expected.trim())) {
-                throw new ComparisonFailure("Text differences", expected, actual);
-            }
+            case STRICT:
+            case STRICT_ANY_ORDER:
+                if (!expected.trim().equals(actual.trim())) {
+                    throw new ComparisonFailure("Text differences", expected, actual);
+                }
+                break;
+            case LOOSE:
+                if (!expected.trim().contains(actual.trim())) {
+                    throw new ComparisonFailure("Text differences", expected, actual);
+                }
         }
     }
 
-
-
-    @Override
-    public <T> void assertFragment(
-        String fragment,
-        ValidatableResponse response,
-        Class<T> dataType,
-        Assertion<T> assertion
-    ) {
-        throw new UnsupportedOperationException("Not implemented for content type "+contentType());
-    }
-
-
-    @Override
-    public void assertContentSchema(String expectedSchema, String content) {
-        throw new UnsupportedOperationException("Not implemented for content type "+contentType());
-    }
 }