+++ kukumo-plugins/kukumo-rest/src/main/java/module-info.java
@@ -27,12 +27,12 @@ module kukumo.rest {
     requires org.hamcrest;
     requires org.everit.json.schema;
     requires java.xml;
+    requires plexus.utils;
 
     uses ConfigContributor;
     uses StepContributor;
 
     provides ConfigContributor with RestConfigContributor;
     provides StepContributor with RestStepContributor;
-
 
 }