+++ kukumo-plugins/kukumo-rest/src/test/java/iti/kukumo/rest/helpers/TestXMLHelper.java
@@ -7,14 +7,13 @@
 /**
  * @author Luis Iñesta Gelabert - linesta@iti.es | luiinge@gmail.com
  */
-package iti.kukumo.rest.test.helpers;
+package iti.kukumo.rest.helpers;
 
 
 import org.junit.ComparisonFailure;
 import org.junit.Test;
 
 import iti.kukumo.rest.MatchMode;
-import iti.kukumo.rest.helpers.XMLHelper;