+++ kukumo-server-quarkus/src/test/java/iti/kukumo/server/ExecutionResourceTest.java
@@ -6,21 +6,21 @@
 
 package iti.kukumo.server;
 
-import static io.restassured.RestAssured.given;
-
-import java.io.IOException;
-import java.nio.file.*;
-import java.util.*;
-
-import org.junit.jupiter.api.Test;
-
 import com.fasterxml.jackson.databind.ObjectMapper;
-
 import io.quarkus.test.junit.QuarkusTest;
 import io.quarkus.test.security.TestSecurity;
+import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
 import io.restassured.response.Response;
+import io.restassured.specification.RequestSpecification;
 import iti.kukumo.api.KukumoConfiguration;
+import org.junit.jupiter.api.Test;
+
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.util.HashMap;
+import java.util.Map;
 
 @QuarkusTest
 @TestSecurity(authorizationEnabled = false)
@@ -33,26 +33,25 @@ class ExecutionResourceTest {
     @Test
     void runKukumoWithSingleFile() throws Exception {
         runKukumoExecutionWithContentOf("simpleScenario.feature").then()
-            .statusCode(OK);
+                .statusCode(OK);
         given()
-            .when()
-            .get("/executions/{executionID}",executionID)
-            .then()
-            .statusCode(200);
+                .when()
+                .get("/executions/{executionID}", executionID)
+                .then()
+                .statusCode(200);
     }
 
 
     @Test
     void runKukumoWithWorkspace() throws Exception {
         runKukumoExecutionWithWorkspace("src/test/resources").then()
-            .statusCode(OK);
+                .statusCode(OK);
         given()
-            .when()
-            .get("/executions/{executionID}",executionID)
-            .then()
-            .statusCode(200);
+                .when()
+                .get("/executions/{executionID}", executionID)
+                .then()
+                .statusCode(200);
     }
-
 
 
     @Test
@@ -61,34 +60,32 @@ class ExecutionResourceTest {
         runKukumoExecutionWithContentOf("simpleScenario.feature");
         runKukumoExecutionWithContentOf("simpleScenario.feature");
         given()
-            .when()
-            .get("/executions?size=1&page=2")
-            .prettyPeek()
-            .then()
-            .statusCode(OK);
+                .when()
+                .get("/executions?size=1&page=2")
+                .prettyPeek()
+                .then()
+                .statusCode(OK);
     }
 
 
     private Response runKukumoExecutionWithContentOf(String feature) throws Exception {
         return extractExecutionId(
-            given()
-            .when()
-            .contentType(ContentType.TEXT)
-            .body(Files.readString(Path.of("src/test/resources/"+feature)))
-            .post("/executions?resourceType=gherkin")
+                given()
+                        .when()
+                        .contentType(ContentType.TEXT)
+                        .body(Files.readString(Path.of("src/test/resources/" + feature)))
+                        .post("/executions?resourceType=gherkin")
         );
     }
 
     private Response runKukumoExecutionWithWorkspace(String workspace) throws Exception {
         return extractExecutionId(
-            given()
-            .when()
-            .contentType(ContentType.TEXT)
-            .post("/executions?workspace="+workspace)
+                given()
+                        .when()
+                        .contentType(ContentType.TEXT)
+                        .post("/executions?workspace=" + workspace)
         );
     }
-
-
 
     private Response extractExecutionId(Response response) throws IOException {
         var data = response.body().print();
@@ -96,8 +93,11 @@ class ExecutionResourceTest {
             return response;
         }
         var json = new ObjectMapper().readValue(data, HashMap.class);
-        this.executionID = (String)((Map<?,?>)json.get("data")).get(KukumoConfiguration.EXECUTION_ID);
+        this.executionID = (String) ((Map<?, ?>) json.get("data")).get(KukumoConfiguration.EXECUTION_ID);
         return response;
     }
 
+    private static RequestSpecification given() {
+        return RestAssured.given().port(8081);
+    }
 }