+++ kukumo-server-quarkus/src/test/java/iti/kukumo/server/KukumoTestSteps.java
@@ -10,24 +10,22 @@
 package iti.kukumo.server;
 
 
-import java.io.*;
-import java.text.ParseException;
-import java.util.Locale;
-
 import iti.kukumo.api.KukumoDataTypeRegistry;
 import iti.kukumo.api.KukumoStepRunContext;
 import iti.kukumo.api.annotations.I18nResource;
 import iti.kukumo.api.annotations.Step;
+import iti.kukumo.api.datatypes.Assertion;
+import iti.kukumo.api.extensions.StepContributor;
 import iti.kukumo.api.plan.DataTable;
 import iti.kukumo.api.plan.Document;
 import org.assertj.core.api.Assertions;
 import org.assertj.core.data.Offset;
 
-import iti.kukumo.api.*;
-import iti.kukumo.api.annotations.*;
-import iti.kukumo.api.datatypes.Assertion;
-import iti.kukumo.api.extensions.StepContributor;
-import iti.kukumo.api.plan.*;
+import java.io.BufferedReader;
+import java.io.IOException;
+import java.io.StringReader;
+import java.text.ParseException;
+import java.util.Locale;
 
 
 @I18nResource("test-kukumo-steps")
@@ -48,7 +46,7 @@ public class KukumoTestSteps implements StepContributor {
     }
 
 
-    @Step(value = "given.two.numbers", args = { "value1:int", "value2:float" })
+    @Step(value = "given.two.numbers", args = {"value1:int", "value2:float"})
     public void setNumbers(Integer value1, Float value2) throws Exception {
         this.value1 = value1;
         this.value2 = value2;
@@ -115,10 +113,10 @@ public class KukumoTestSteps implements StepContributor {
 
 
     @Step(value = "simple.step.with.multiple.asserts", args = {
-		"a:integer-assertion",
-		"b:integer",
-        "c:text-assertion"
-	})
+            "a:integer-assertion",
+            "b:integer",
+            "c:text-assertion"
+    })
     public void simpleStepWithMultipleAsserts(Assertion<Integer> a, Long b, Assertion<String> c) {
         // nothing
     }